Reverse change, found breakage cause
Again, sorry for the spam of commits, but I have found that apparently changing the .md file name causes the post to disappear from the feed on the github.io Jekyll website. Not sure why, as I thought it relies on the title in the front matter to generate the link and such, not the .md file name. Thankfully the existing date format is still usable going forward.